manager/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "manager"
authors = [
  { name = "Kiwix", email = "dev@kiwix.org" },
]
keywords = ["kiwix", "imager", "manager"]
requires-python = ">=3.11"
description = "Kiwix Imager Service UI"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
dependencies = [
  "Django>=2.2.11,<2.3",
  "humanfriendly==10.0",
  "langcodes==1.4.1",
  "python-magic==0.4.15",
  "PyYAML==6.0.1",
  "requests==2.31.0",
  "jsonfield==2.0.2",
  "pycountry==18.5.26",
  "phonenumbers==8.9.15",
  "python-dateutil==2.7.5",
  "ansi2html==1.5.2",
  "mysqlclient==2.2.0",
  "werkzeug>=1.0.1",
  "Babel==2.9.1",
  # jinja2 dependency (https://github.com/pallets/markupsafe/issues/284)
  "MarkupSafe==2.0.1",
  "offspot-config==1.3.0",
  "Pillow==10.0.0",
  "xmltodict==0.13.0",

]
version = "2.0.0.dev0"

[project.optional-dependencies]
scripts = [
  "invoke==2.2.0",
]
lint = [
  "black==23.9.1",
  "ruff==0.0.291",
]
check = [
  "pyright==1.1.328",
]
test = [
  "pytest==7.4.2",
  "coverage==7.3.1",
]
dev = [
  "pre-commit==3.4.0",
  "debugpy==1.8.0",
  "ipython==8.15.0",
  "manager[scripts]",
  "manager[lint]",
  "manager[test]",
  "manager[check]",
]

[project.urls]
Homepage = "https://github.com/offspot/cardshop"
Donate = "https://www.kiwix.org/en/support-us/"

[tool.hatch.build]
exclude = [
  "/.github",
]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.test]
features = ["scripts", "test"]

[[tool.hatch.envs.test.matrix]]
python = ["3.11"]

[tool.hatch.envs.test.scripts]
run = "inv test --args '{args}'"
run-cov = "inv test-cov --args '{args}'"
report-cov = "inv report-cov"
coverage = "inv coverage --args '{args}'"
html = "inv coverage --html --args '{args}'"

[tool.hatch.envs.lint]
template = "lint"
skip-install = false
features = ["scripts", "lint"]

[tool.hatch.envs.lint.scripts]
black = "inv lint-black --args '{args}'"
ruff = "inv lint-ruff --args '{args}'"
all = "inv lintall --args '{args}'"
fix-black = "inv fix-black --args '{args}'"
fix-ruff = "inv fix-ruff --args '{args}'"
fixall = "inv fixall --args '{args}'"

[tool.hatch.envs.check]
features = ["scripts", "check"]

[tool.hatch.envs.check.scripts]
pyright = "inv check-pyright --args '{args}'"
all = "inv checkall --args '{args}'"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]
select = [
  "A",  # flake8-builtins
  # "ANN",  # flake8-annotations
  "ARG",  # flake8-unused-arguments
  # "ASYNC",  # flake8-async
  "B",  # flake8-bugbear
  # "BLE",  # flake8-blind-except
  "C4",  # flake8-comprehensions
  "C90",  # mccabe
  # "COM",  # flake8-commas
  # "D",  # pydocstyle
  # "DJ",  # flake8-django
  # "DTZ",  # flake8-datetimez
  "E",  # pycodestyle (default)
  "EM",  # flake8-errmsg
  # "ERA",  # eradicate
  # "EXE",  # flake8-executable
  "F",  # Pyflakes (default)
  # "FA",  # flake8-future-annotations
  "FBT",  # flake8-boolean-trap
  # "FLY",  # flynt
  # "G",  # flake8-logging-format
  "I",  # isort
  "ICN",  # flake8-import-conventions
  # "INP",  # flake8-no-pep420
  # "INT",  # flake8-gettext
  # "ISC",  # flake8-implicit-str-concat
  "N",  # pep8-naming
  # "NPY",  # NumPy-specific rules
  # "PD",  # pandas-vet
  # "PGH",  # pygrep-hooks
  # "PIE",  # flake8-pie
  # "PL",  # Pylint
  "PLC",  # Pylint: Convention
  "PLE",  # Pylint: Error
  "PLR",  # Pylint: Refactor
  "PLW",  # Pylint: Warning
  # "PT",  # flake8-pytest-style
  # "PTH",  # flake8-use-pathlib
  # "PYI",  # flake8-pyi
  "Q",  # flake8-quotes
  # "RET",  # flake8-return
  # "RSE",  # flake8-raise
  "RUF",  # Ruff-specific rules
  "S",  # flake8-bandit
  # "SIM",  # flake8-simplify
  # "SLF",  # flake8-self
  "T10",  # flake8-debugger
  "T20",  # flake8-print
  # "TCH",  # flake8-type-checking
  # "TD",  # flake8-todos
  "TID",  # flake8-tidy-imports
  # "TRY",  # tryceratops
  "UP",  # pyupgrade
  "W",  # pycodestyle
  "YTT",  # flake8-2020
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Remove flake8-errmsg since we consider they bloat the code and provide limited value
  "EM",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore warnings on subprocess.run / popen
  "S603",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["manager"]

[tool.ruff.flake8-bugbear]
# add exceptions to B008 for fastapi.
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
"**/convertdata.py" = ["E501"]

[tool.pytest.ini_options]
minversion = "7.3"
testpaths = ["tests"]
pythonpath = [".", "manager"]

[tool.coverage.paths]
great_project = ["manager"]
tests = ["tests"]

[tool.coverage.run]
source_pkgs = ["manager"]
branch = true
parallel = true
omit = [
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pyright]
include = ["manager", "tests", "tasks.py"]
exclude = [".env/**", ".venv/**"]
extraPaths = ["manager"]
pythonVersion = "3.11"
typeCheckingMode="basic"
